% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_result_f}
\alias{check_result_f}
\title{Checks that user code and solution code results are equal.}
\usage{
check_result_f(
  label,
  user_code,
  solution_code,
  check_code,
  envir_result,
  evaluate_result,
  envir_prep,
  ...
)
}
\arguments{
\item{label}{Label for exercise chunk.}

\item{user_code}{R code submitted by the user.}

\item{solution_code}{Code provided within the “-solution” chunk for the exercise.}

\item{check_code}{Code provided within the “-check” chunk for the exercise.}

\item{envir_result}{The R environment after the execution of the chunk.}

\item{evaluate_result}{The return value from the evaluate::evaluate function.}

\item{envir_prep}{A copy of the R environment before the execution of the chunk.}

\item{...}{Additional parameters}
}
\value{
list with message
}
\description{
Checks that user code and solution code results are equal.
}
\examples{
check_result_f("ex1", "x <- 5", "x <- 5", "TRUE")
}
