% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_csv_path.R
\name{example_csv_path}
\alias{example_csv_path}
\title{Helper function that gets the path of the example csv file included
with BeTidy}
\usage{
example_csv_path()
}
\value{
String with path
}
\description{
Helper function that gets the path of the example csv file included
with BeTidy
}
\examples{
example_csv_path()
}
