% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_paths.R
\name{example_tsv_path}
\alias{example_tsv_path}
\title{Helper function that gets the path of the example tsv file included
with BeTidy}
\usage{
example_tsv_path()
}
\value{
String with path
}
\description{
Helper function that gets the path of the example tsv file included
with BeTidy
}
\examples{
example_tsv_path()
}
